requirements.txt
asgiref==3.8.1
certifi==2024.7.4
cffi==1.16.0
channels==4.1.0
charset-normalizer==3.3.2
cloudinary==1.40.0
cryptography==42.0.8
Django
django-allauth==0.63.3
django-cloudinary-storage==0.3.0
django-jazzmin==3.0.0
django-rest-auth==0.9.5
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
idna==3.7
inflection==0.5.1
jwt==1.3.1
packaging==24.1
pillow==10.4.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-decouple==3.8
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.2.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
